contest.c 원본 코드에 반영. 원본코드 테스트 성공. 테스트 벡터와 벤치마크만 남음.
